Rename view model
